Differentiate available actions on creating a post vs comment
