+++ stream-feeds-android-sample/src/main/java/io/getstream/feeds/android/sample/feed/CommentsBottomSheet.kt
@@ -160,14 +160,13 @@ private fun ColumnScope.CommentsBottomSheetContent(
     }
 
     if (createCommentData != null) {
-        // TODO [G.] differentiate between creating a post vs a comment, e.g. no polls in comments
         CreateContentBottomSheet(
+            title = "Add comment",
             onDismiss = { createCommentData = null },
             onPost = { text, attachments ->
                 onPostComment(text, createCommentData?.replyParentId, attachments)
                 createCommentData = null
             },
-            onCreatePoll = {},
         )
     }
 }
